Move positionsHistory sync-schema to separate file
